make it work with sass
accoutrement-color doesn't fully work because it depends on sass
features that aren't present in node-sass, especially first-class
functions. So this doesn't work:

    border-color: color('md-grey' ('lighten': 10%));

because 'lighten' can't be referenced by name. It's hacked for now to
this instead:

    border-color: color('md-grey'); // XXX ('lighten': 10%));

Also the original sass depended on susy v2 but the current version is
v3. Since we're not using the susy grid in the markup, just use
a simplified container to get the build running.