use our own gatsby-plugin-adobe-fonts
The problem with gatsby-plugin-web-font-loader is that it doesn't build
as part of SSR, instead it runs dynamically after page load, so it takes
a long time to load fonts.

This plugin works the same way as gatsby-plugin-google-analytics, by
installing the script snippet directly into the SSR page. Additionally,
we hide the body until the fonts have loaded (or failed to load) to
avoid FOUT.